Remove unused model properties and simplify API mapper
