+++ models/index.ts
@@ -1,4 +1,3 @@
 export type { Nutriments } from './nutriments';
-export type { NutrientLevels } from './nutrient-levels';
 export type { Product, ProductResult } from './product';
 export type { MealCategory, MealItem, DayLog, NutritionGoals } from './meal-entry';
