--- models/nutrient-levels.ts
@@ -1,6 +1,0 @@
-export interface NutrientLevels {
-  fat: string | null;
-  saturatedFat: string | null;
-  sugars: string | null;
-  salt: string | null;
-}
